requirements.txt
discord.py==2.4.0
gTTS==2.5.3
python-dotenv==1.0.1
yt_dlp==2024.11.4
